Commit cmake build system
Added build system to the project. Currently manages dependencies
for: rapidjson, lua 5.3, sdl2, sfml and googleTest.

The engine currently consists of two dummy source files, this is
so that CMake does not complain about the engine not having any
source files.

Please note: for the project/sound/ project to work, create a
folder "project/sound/assets/" and place a file called "sound.wav"
into it. Then build/rebuild the project.
I still haven't figured out how we should handle asset loading.
This is just a quick and easy way to do it.