Fill-out keyboard functionality, change Keyboard namespace
Added a number of recognized keys to the Keyboard::Key enumeration,
so that we can use those keys. Added 'PrintKeys' function that
should be used to print keys (allows us to print non-printable
characters such as return or caps lock).

The namespace of the Keyboard class was changed from ds to
ds_platform to be more consistent with the other low-level classes.