Add TextureResource Class
Add the TextureResource class that handles the
reading of texture image files. Required a
filepath be passed into it's static function.

Modify the individual resource files to take
in a file path, instead of a fstream. Some
of the used libs don't require an fstream be
passed in and the fstream doesn't provide
the file name.